Use float as positions for the camera

+++ include/camera.h
@@ -4,10 +4,10 @@
 #define _CAMERA_H
 
 struct {
-  i32 x;
-  i32 y;
-  i32 x_target;
-  i32 y_target;
+  float x;
+  float y;
+  float x_target;
+  float y_target;
   struct Entity* target;
   u8 has_target;
 } camera;
